update template, crud brand, crud invoice, online payment

+++ src/main/java/com/example/mpds/services/impl/ProductService.java
@@ -2,11 +2,11 @@ package com.example.mpds.services.impl;
 
 import com.example.mpds.dto.ProductDTO;
 import com.example.mpds.entity.CategoryEntity;
+import com.example.mpds.entity.InvoiceInfoEntity;
 import com.example.mpds.entity.ProductEntity;
+import com.example.mpds.entity.ProductReviewEntity;
 import com.example.mpds.mapper.ProductMapper;
-import com.example.mpds.repository.CategoryRepository;
-import com.example.mpds.repository.ProductRepository;
-import com.example.mpds.repository.ProductSpecifications;
+import com.example.mpds.repository.*;
 import com.example.mpds.services.IProductService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -16,6 +16,7 @@ import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 public class ProductService implements IProductService {
@@ -26,37 +27,15 @@ public class ProductService implements IProductService {
     private CategoryRepository categoryRepository;
     @Autowired
     private ProductMapper mapper;
+    @Autowired
+    private ProductReviewRepository productReviewRepository;
+
+    @Autowired
+    private InvoiceInfoRepository invoiceInfoRepository;
 
     @Override
     public FilterProductResult findAll(List<Integer> categories, List<String> types, List<String> dialSizes, Pageable pageable, String searchText) {
-//        List<ProductDTO> results=new ArrayList<>();
-//        List<ProductEntity> productList=productRepository.findAll(pageable).getContent();
-//        for(ProductEntity item:productList){
-//            ProductDTO productDTO=mapper.toDTO(item);
-//            results.add(productDTO);
-//        }
-//        return results;
-//        if (categories == null || categories.isEmpty()) {
-//            categories = List.of(""); // Assign default value for categories
-//        }
-//        if (types == null || types.isEmpty()) {
-//            types = List.of(""); // Assign default value for types
-//        }
-//        if (dialSizes == null || dialSizes.isEmpty()) {
-//            dialSizes = List.of(""); // Assign default value for dialSizes
-//        }
-//
-//        if (categories.size() == 1 && categories.get(0).isEmpty() &&
-//                types.size() == 1 && types.get(0).isEmpty() &&
-//                dialSizes.size() == 1 && dialSizes.get(0).isEmpty()) {
-//            // No filters provided, retrieve all products
-//            productEntities = productRepository.findAll(pageable).getContent();
-//        } else {
-//            // Filters provided, apply them
-//
-//        }
-//        String typesAsString = types != null ? String.join(",", types) : null;
-//        String dialSizesAsString = dialSizes != null ? String.join(",", dialSizes) : null;
+
 
         Page<ProductEntity> productEntities;
 
@@ -79,56 +58,75 @@ public class ProductService implements IProductService {
     }
 
     public List<ProductDTO> findAll() {
-        List<ProductDTO> results=new ArrayList<>();
-        List<ProductEntity> productList=productRepository.findAll();
-        for(ProductEntity item:productList){
-            ProductDTO productDTO=mapper.toDTO(item);
+        List<ProductDTO> results = new ArrayList<>();
+        List<ProductEntity> productList = productRepository.findAll();
+        for (ProductEntity item : productList) {
+            ProductDTO productDTO = mapper.toDTO(item);
             System.out.println(productDTO.getCategoryName());
             results.add(productDTO);
         }
         return results;
     }
-    public ProductDTO findOne(String name){
-        ProductDTO productDTO=new ProductDTO();
-        ProductEntity product=productRepository.findOneByName(name);
-        productDTO= mapper.toDTO(product);
+
+    public ProductDTO findOne(String name) {
+        ProductDTO productDTO = new ProductDTO();
+        ProductEntity product = productRepository.findOneByName(name);
+        productDTO = mapper.toDTO(product);
         return productDTO;
     }
 
-    public ProductDTO findOne(int id){
-        ProductDTO productDTO=new ProductDTO();
-        ProductEntity product=productRepository.findOneById(id);
-        productDTO= mapper.toDTO(product);
+    public ProductDTO findOne(int id) {
+        ProductDTO productDTO = new ProductDTO();
+        ProductEntity product = productRepository.findOneById(id);
+        productDTO = mapper.toDTO(product);
         return productDTO;
     }
-    public ProductEntity findProduct(int id){
 
-        ProductEntity product=productRepository.findOneById(id);
+    public ProductEntity findProduct(int id) {
+
+        ProductEntity product = productRepository.findOneById(id);
 
         return product;
     }
-    public  void insertProduct(ProductDTO product){
+
+    public String insertProduct(ProductDTO product) {
         ProductEntity newProd = mapper.toEntity(product);
         CategoryEntity category = categoryRepository.findByName(product.getCategoryName());
-        newProd.setCategory(category);
-        productRepository.save(newProd);
+        if (category != null) {
+            newProd.setCategory(category);
+            productRepository.save(newProd);
+            return "Insert Product Suscessfully";
+        } else return "Brand name: " + product.getCategoryName() + " does not exist!";
     }
-    public int totalProduct(){
+
+    public int totalProduct() {
         return (int) productRepository.count();
     }
 
-    public void deleteById(Long id){
-        productRepository.deleteById(id);
+    public String deleteById(Long id) {
+        List<InvoiceInfoEntity> lstInvoiceInfor = invoiceInfoRepository.findByProduct_Id(id.intValue());
+        if (lstInvoiceInfor.isEmpty()) {
+            productRepository.deleteById(id);
+            return "Delete Suscessfully";
+        }
+
+        List<Integer> invoiceIds = lstInvoiceInfor.stream()
+                .map(invoiceInfo -> invoiceInfo.getInvoice().getId())
+                .collect(Collectors.toList());
+
+        return "Delete Failed, Invoices associated with product ID " + id + ": " + invoiceIds;
     }
 
-    public void updateProduct(ProductDTO product)
-    {
+    public String updateProduct(ProductDTO product) {
 
         ProductEntity existingProduct = productRepository.findById((long) product.getId())
                 .orElseThrow(() -> new IllegalArgumentException("Product not found"));
 
-
         CategoryEntity category = categoryRepository.findByName(product.getCategoryName());
+        if (category == null){
+            return "Update Failed! Brand "+product.getCategoryName()+" does not exist!";
+        }
+
         existingProduct.setCategory(category);
         existingProduct.setName(product.getName());
         existingProduct.setPrice(product.getPrice());
@@ -137,23 +135,11 @@ public class ProductService implements IProductService {
         existingProduct.setImage3(product.getImage3());
         existingProduct.setDescription(product.getDescription());
         productRepository.save(existingProduct);
+        return "Update Suscessfully";
     }
 
-//    public List<ProductDTO> filterProduct(List<String> categories, List<String> types, List<String> dialSizes, Pageable pageable) {
-//        if (categories == null || categories.isEmpty()) {
-//            categories = List.of(""); // Assign default value for categories
-//        }
-//        if (types == null || types.isEmpty()) {
-//            types = List.of(""); // Assign default value for types
-//        }
-//        if (dialSizes == null || dialSizes.isEmpty()) {
-//            dialSizes = List.of(""); // Assign default value for dialSizes
-//        }
-//        List<ProductEntity> productEntities = productRepository.findByCategoriesAndTypesAndDialSizes(categories, types, dialSizes, pageable);
-//        List<ProductDTO> productDTOs = new ArrayList<>();
-//        for (ProductEntity entity : productEntities) {
-//            productDTOs.add(mapper.toDTO(entity));
-//        }
-//        return productDTOs;
-//    }
+    public List<ProductReviewEntity> getReviewsByProductId(int productId) {
+        return productReviewRepository.findByProductId(productId);
+    }
+
 }
